feat: extract test case into separate file
